Entregables/Leopoldo Prado Perez/Demobehave/requirements.txt
behave
Appium-Python-Client
allure-behave==2.13.2
python-dotenv

h11
pip
wheel
trio
outcome
attrs
sniffio
exceptiongroup
sortedcontainers
cffi
idna
selenium
urllib3
six
parse
pluggy
certifi
pycparser
setuptools
wsproto